Move hooks to VoilaConfiguration
